requirements.txt
Django==2.1.7
django-environ==0.4.5
django-split-settings==0.3.0
django-webpack-loader==0.6.0
psycopg2==2.7.5
pytz==2018.5
raven==6.9.0
